+++ CHANGELOG.md
@@ -14,9 +14,9 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 - Support 64-bit atomics on ARMv7-R and ARMv8-R.
 
-- Fix bug in AArch64 128-bit SeqCst load when FEAT_LSE2 is enabled at compile-time.
-
 - Fix bug in AArch64 16-bit compare_exchange_weak.
+
+- Fix bug in AArch64 128-bit SeqCst load when FEAT_LSE2 is enabled at compile-time.
 
 - Optimize AArch64 atomic load when the `rcpc` target feature is enabled at compile-time.
 
@@ -26,7 +26,7 @@ Note: In this file, do not use the hard wrap in the middle of a sentence for com
 
 **Note:** This release has been yanked due to a bug fixed in 0.2.16.
 
-- Improve support for AArch32 ARMv8.
+- Improve support for ARMv8 AArch32.
 
 - Optimize inline assemblies on x86.